backend/requirements.txt
# SageFlow Backend Requirements
# Python 3.11+

# FastAPI Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Redis & Celery
redis==5.0.4
celery==5.4.0

# LangChain & RAG
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0

# Vector Database
qdrant-client==1.9.0

# Embeddings
sentence-transformers==3.0.0

# Document Processing
pymupdf==1.24.3
python-docx==1.1.0
python-multipart==0.0.9

# HTTP Client
httpx==0.27.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email Validation
email-validator==2.1.1

# System Monitoring
psutil==5.9.8
